vindex.h: result and remove refactor
_IResult and _IConstResult interfaces have been stabilized and
massively simplified. There is no need to even have a pure
virtual data() setter. Just make the getter return a reference or
const reference. Also, for _IResults, use universal/forwarding
references to pass by reference up until, and possibly, the setting
of the instance member. It's a "possibly" depending on if T is a
reference or a non-reference

For remove(), the logic between removing from the _insertion_list
and the _index versus removing from the AVL tree has been decoupled.
Also, remove() now returns a Result<T> now so the user can detect
if a removal resulted in a no-op or not